+++ ReleaseNotes.md
@@ -1,3 +1,7 @@
+### New in 0.6.5 (Released 2018/08/12)
+* [Improvement] Not copying cache-control
+* [Improvement] Add web-fonts to default compress extensions
+
 ### New in 0.6.4 (Released 2018/08/12)
 * [Improvement] Updated Build Scripts